+++ lib/mongoid/document.rb
@@ -242,7 +242,7 @@ module Mongoid #:nodoc:
         allocate.tap do |doc|
           doc.instance_variable_set(:@attributes, attributes)
           doc.send(:apply_defaults)
-          IdentityMap.set(doc)
+          IdentityMap.set(doc) unless _loading_revision?
           doc.run_callbacks(:initialize) { doc }
         end
       end