+++ lib/mongoid/threaded/lifecycle.rb
@@ -139,6 +139,23 @@ module Mongoid #:nodoc:
         Threaded.loading?
       end
 
+      # Execute a block in loading revision mode.
+      #
+      # @example Execute in loading revision mode.
+      #   _loading_revision do
+      #     load_revision
+      #   end
+      #
+      # @return [ Object ] The return value of the block.
+      #
+      # @since 2.3.4
+      def _loading_revision
+        Threaded.begin_load_revision
+        yield
+      ensure
+        Threaded.exit_load_revision
+      end
+
       module ClassMethods #:nodoc:
 
         # Execute a block in creating mode.
@@ -156,6 +173,18 @@ module Mongoid #:nodoc:
           yield
         ensure
           Threaded.exit_create
+        end
+
+        # Is the current thread in loading revision mode?
+        #
+        # @example Is the current thread in loading revision mode?
+        #   proxy._loading_revision?
+        #
+        # @return [ true, false ] If the thread is loading a revision.
+        #
+        # @since 2.3.4
+        def _loading_revision?
+          Threaded.loading_revision?
         end
       end
     end